Correction bug dans la transactiion

+++ app/api/transaction/nature/[id]/route.ts
@@ -57,6 +57,39 @@ export async function DELETE(req: NextRequest) {
 
     const nature = await prisma.transactionNature.findUnique({
         where: { id },
+        include: {
+            receipts: true,
+            dibursements: true,
+            userActions: true
+        }
+    });
+
+    if (nature?.receipts && nature.receipts.length > 0) {
+        return NextResponse.json({
+            state: "error",
+            message: "Impossible de supprimer cette nature car des encaissements sont liés à cette nature.",
+        }, { status: 409 });
+    }
+
+    if (nature?.dibursements && nature.dibursements.length > 0) {
+        return NextResponse.json({
+            state: "error",
+            message: "Impossible de supprimer cette nature car des décaissements sont liés à cette nature.",
+        }, { status: 409 });
+
+    }
+
+    if (nature?.userActions && nature.userActions.length > 0) {
+        return NextResponse.json({
+            state: "error",
+            message: "Impossible de supprimer cette nature car des client | fournisseurs | tiers sont liés à cette nature.",
+        }, { status: 409 });
+
+    }
+
+
+    await prisma.transactionNature.findUnique({
+        where: { id },
     });
 
     const deletedNature = await prisma.transactionNature.delete({ where: { id } });
